페인터 기능 추가

+++ etc/edit/src/components/Map/Editor/Aside/Item/Terrain/handleEditApply.js
@@ -22,6 +22,12 @@ export default ({
     indexes.push(i);
   });
 
+  if (_.isEmpty(indexes)) {
+    dispatch(set({ editable: undefined }));
+
+    return;
+  }
+
   axios.delete(`/terrains/${data.id}/sprite/${_.join(indexes)}`);
 
   dispatch(set({
